wallet error module
